refactor general
