+++ config.ini
@@ -1,0 +1,28 @@
+#formato de configuracion: tipo de campo=letra de columna
+[Iva Compras]
+punto de ventas=C
+numero de comprobante=D
+tipo de comprobante=B
+denominacion=H
+numero de documento=G
+tipo de documento=F
+tipo de cambio=I
+neto=K
+neto no gravado=L
+exento=M
+iva=N
+total=O
+
+[Iva Ventas]
+punto de ventas=C
+numero de comprobante=D
+tipo de comprobante=B
+denominacion=H
+numero de documento=G
+tipo de documento=F
+tipo de cambio=I
+neto=K
+neto no gravado=L
+exento=M
+iva=N
+total=O
